+++ packages/ui/src/form/primitives/autocomplete/popup.tsx
@@ -4,17 +4,25 @@ import { forwardRef } from '@wordpress/element';
 import { renderSlotWithChildren } from '../../../utils/render-slot-with-children';
 import { ThemeProvider } from '../../../utils/theme-provider';
 import itemPopupStyles from '../../../utils/css/item-popup.module.css';
+import { getItemPopupWidthClassName } from '../../../utils/css/item-popup';
 import { Portal } from './portal';
 import { Positioner } from './positioner';
 import type { AutocompletePopupProps } from './types';
 
 export const Popup = forwardRef< HTMLDivElement, AutocompletePopupProps >(
-	function Popup( { className, portal, positioner, ...restProps }, ref ) {
+	function Popup(
+		{ className, portal, positioner, popupWidth, ...restProps },
+		ref
+	) {
 		const popupContent = (
 			<ThemeProvider>
 				<_Autocomplete.Popup
 					ref={ ref }
-					className={ clsx( itemPopupStyles.popup, className ) }
+					className={ clsx(
+						itemPopupStyles.popup,
+						getItemPopupWidthClassName( popupWidth ),
+						className
+					) }
 					{ ...restProps }
 				/>
 			</ThemeProvider>