+++ packages/ui/src/form/primitives/select/popup.tsx
@@ -6,18 +6,23 @@ import { Positioner } from './positioner';
 import { renderSlotWithChildren } from '../../../utils/render-slot-with-children';
 import { ThemeProvider } from '../../../utils/theme-provider';
 import itemPopupStyles from '../../../utils/css/item-popup.module.css';
+import { getItemPopupWidthClassName } from '../../../utils/css/item-popup';
 import type { SelectPopupProps } from './types';
 
 export const Popup = forwardRef< HTMLDivElement, SelectPopupProps >(
 	function Popup(
-		{ className, portal, positioner, children, ...restProps },
+		{ className, portal, positioner, popupWidth, children, ...restProps },
 		ref
 	) {
 		const popupContent = (
 			<ThemeProvider>
 				<_Select.Popup
 					ref={ ref }
-					className={ clsx( itemPopupStyles.popup, className ) }
+					className={ clsx(
+						itemPopupStyles.popup,
+						getItemPopupWidthClassName( popupWidth ),
+						className
+					) }
 					{ ...restProps }
 				>
 					<_Select.List className={ itemPopupStyles.list }>